Use bit field for struct data_object to reduce memory footprint
Reduces the required RAM and flash memory by 4 bytes * number of data objects.

Drawback: Only 8 instead of 16 different subsets can be defined (which is
considered plenty) and the maximum length of strings (specified in "detail"
struct member) is 2^11.

+++ src/thingset.c
@@ -65,7 +65,7 @@ int ts_process(struct ts_context *ts, const uint8_t *request, size_t request_len
     }
 }
 
-void ts_set_authentication(struct ts_context *ts, uint16_t flags)
+void ts_set_authentication(struct ts_context *ts, uint8_t flags)
 {
     ts->_auth_flags = flags;
 }
