+++ src/thingset_bin.c
@@ -345,7 +345,7 @@ int ts_bin_fetch(struct ts_context *ts, const struct ts_data_object *parent, uin
     }
 }
 
-int ts_bin_import(struct ts_context *ts, uint8_t *data, size_t len, uint16_t auth_flags,
+int ts_bin_import(struct ts_context *ts, uint8_t *data, size_t len, uint8_t auth_flags,
                   uint16_t subsets)
 {
     uint8_t resp_tmp[1] = {};   // only one character as response expected
@@ -358,7 +358,7 @@ int ts_bin_import(struct ts_context *ts, uint8_t *data, size_t len, uint16_t aut
 }
 
 int ts_bin_patch(struct ts_context *ts, const struct ts_data_object *parent,
-                 unsigned int pos_payload, uint16_t auth_flags, uint16_t subsets)
+                 unsigned int pos_payload, uint8_t auth_flags, uint16_t subsets)
 {
     unsigned int pos_req = pos_payload;
     uint16_t num_elements, element = 0;
